Rename all "InputManager" to "InputContainer"
